:shirt: Move interactions to shared library
